+++ astrocut/footprint_cutout.py
@@ -10,6 +10,7 @@ from astropy.coordinates import SkyCoord
 from astropy.table import Table, Column
 from cachetools import TTLCache, cached
 from spherical_geometry.polygon import SphericalPolygon
+from spherical_geometry.vector import radec_to_vector
 
 from .cutout import Cutout
 
@@ -193,26 +194,54 @@ def get_ffis(s3_footprint_cache: str) -> Table:
     return ffis
 
 
-def ra_dec_crossmatch(all_ffis: Table, coordinates: SkyCoord, cutout_size, arcsec_per_px: int = 21) -> Table:
+def _crossmatch_point(ra: SkyCoord, dec: SkyCoord, all_ffis: Table) -> np.ndarray:
     """
-    Returns the Full Frame Images (FFIs) whose footprints overlap with a cutout of a given position and size.
+    Returns the indices of the Full Frame Images (FFIs) that contain the given RA and
+    Dec coordinates by checking which FFI polygons contain the point.
+    
+    Parameters
+    ----------
+    ra : SkyCoord
+        Right Ascension in degrees.
+    dec : SkyCoord
+        Declination in degrees.
+    all_ffis : `~astropy.table.Table`
+        Table of FFIs to crossmatch with the point.
+        
+    Returns
+    -------
+    ffi_inds : `~numpy.ndarray`
+        Indices of FFIs that contain the given RA and Dec coordinates.
+    """
+    ffi_inds = []
+    vector_coord = radec_to_vector(ra, dec)
+    for sector in np.unique(all_ffis['sequence_number']):
+        # Returns a 2-long array where the first element is indexes and the 2nd element is empty
+        sector_ffi_inds = np.where(all_ffis['sequence_number'] == sector)[0]
+
+        for ind in sector_ffi_inds:
+            if all_ffis[ind]["polygon"].contains_point(vector_coord):
+                ffi_inds.append(ind)
+                break  # the ra/dec will only be on one ccd per sector
+    return np.array(ffi_inds, dtype=int)
+
+
+def _crossmatch_polygon(ra: SkyCoord, dec: SkyCoord, all_ffis: Table, px_size: np.ndarray,
+                        arcsec_per_px: int = 21) -> np.ndarray:
+    """
+    Returns the indices of the Full Frame Images (FFIs) that intersect with the given cutout footprint
+    by checking which FFI polygons intersect with the cutout polygon.
 
     Parameters
     ----------
+    ra : SkyCoord
+        Right Ascension in degrees.
+    dec : SkyCoord
+        Declination in degrees.
     all_ffis : `~astropy.table.Table`
-        Table of FFIs to crossmatch with the cutout.
-    coordinates : str or `astropy.coordinates.SkyCoord` object
-        The position around which to cutout.
-        It may be specified as a string ("ra dec" in degrees)
-        or as the appropriate `~astropy.coordinates.SkyCoord` object.
-    cutout_size : int, array-like, `~astropy.units.Quantity`
-        The size of the cutout array. If ``cutout_size``
-        is a scalar number or a scalar `~astropy.units.Quantity`,
-        then a square cutout of ``cutout_size`` will be created.  If
-        ``cutout_size`` has two elements, they should be in ``(ny, nx)``
-        order.  Scalar numbers in ``cutout_size`` are assumed to be in
-        units of pixels. `~astropy.units.Quantity` objects must be in pixel or
-        angular units.
+        Table of FFIs to crossmatch with the point.
+    px_size : array-like
+        Size of the cutout in pixels, in the form [ny, nx].
     arcsec_per_px : int, optional
         Default 21. The number of arcseconds per pixel in an image. Used to determine
         the footprint of the cutout. Default is the number of arcseconds per pixel in
@@ -220,21 +249,15 @@ def ra_dec_crossmatch(all_ffis: Table, coordinates: SkyCoord, cutout_size, arcse
 
     Returns
     -------
-    matching_ffis : `~astropy.table.Table`
-        Table containing information about FFIs whose footprints overlap those of the cutout.
+    ffi_inds : `~numpy.ndarray`
+        Boolean array indicating whether each FFI intersects with the cutout.
     """
-    # Convert coordinates to SkyCoord
-    if not isinstance(coordinates, SkyCoord):
-        coordinates = SkyCoord(coordinates, unit='deg')
-    ra, dec = coordinates.ra, coordinates.dec
-
-    # Parse cutout size
-    cutout_size = Cutout._parse_size_input(cutout_size)
-
     # Create polygon for intersection
     # Convert dimensions from pixels to arcseconds and divide by 2 to get offset from center
-    ra_offset = ((cutout_size[0] * arcsec_per_px) / 2) * u.arcsec
-    dec_offset = ((cutout_size[1] * arcsec_per_px) / 2) * u.arcsec
+    # If one of the dimensions is 0, use a very small value to avoid issues with SphericalPolygon
+    min_offset = 0.1  # pixels
+    ra_offset = ((max(px_size[0], min_offset) * arcsec_per_px) / 2) * u.arcsec
+    dec_offset = ((max(px_size[1], min_offset) * arcsec_per_px) / 2) * u.arcsec
 
     # Calculate RA and Dec boundaries
     ra_bounds = [ra - ra_offset, ra + ra_offset]
@@ -250,5 +273,66 @@ def ra_dec_crossmatch(all_ffis: Table, coordinates: SkyCoord, cutout_size, arcse
     # Find indices of FFIs that intersect with the cutout
     ffi_inds = np.vectorize(lambda ffi: ffi.intersects_poly(cutout_fp))(all_ffis['polygon'])
     ffi_inds = FootprintCutout._ffi_intersect(all_ffis, cutout_fp)
+
+    return ffi_inds
+
+
+def ra_dec_crossmatch(all_ffis: Table, coordinates: Union[SkyCoord, str], cutout_size, 
+                      arcsec_per_px: int = 21) -> Table:
+    """
+    Returns the Full Frame Images (FFIs) whose footprints overlap with a cutout of a given position and size.
+
+    Parameters
+    ----------
+    all_ffis : `~astropy.table.Table`
+        Table of FFIs to crossmatch with the cutout.
+    coordinates : str or `astropy.coordinates.SkyCoord` object
+        The position around which to cutout.
+        It may be specified as a string ("ra dec" in degrees)
+        or as the appropriate `~astropy.coordinates.SkyCoord` object.
+    cutout_size : int, array-like, `~astropy.units.Quantity`
+        The size of the cutout array. If ``cutout_size``
+        is a scalar number or a scalar `~astropy.units.Quantity`,
+        then a square cutout of ``cutout_size`` will be used.  If
+        ``cutout_size`` has two elements, they should be in ``(ny, nx)``
+        order.  Scalar numbers in ``cutout_size`` are assumed to be in
+        units of pixels. `~astropy.units.Quantity` objects must be in pixel or
+        angular units.
+
+        If a cutout size of zero is provided, the function will return FFIs that contain 
+        the exact RA and Dec position. If a non-zero cutout size is provided, the function 
+        will return FFIs whose footprints overlap with the cutout area.
+    arcsec_per_px : int, optional
+        Default 21. The number of arcseconds per pixel in an image. Used to determine
+        the footprint of the cutout. Default is the number of arcseconds per pixel in
+        a TESS image.
+
+    Returns
+    -------
+    matching_ffis : `~astropy.table.Table`
+        Table containing information about FFIs whose footprints overlap those of the cutout.
+    """
+    # Convert coordinates to SkyCoord
+    if not isinstance(coordinates, SkyCoord):
+        coordinates = SkyCoord(coordinates, unit='deg')
+    ra, dec = coordinates.ra, coordinates.dec
+
+    px_size = np.zeros(2, dtype=object)
+    for axis, size in enumerate(Cutout.parse_size_input(cutout_size, allow_zero=True)):
+        if isinstance(size, u.Quantity):  # If Quantity, convert to pixels
+            if size.unit == u.pixel:
+                px_size[axis] = size.value
+            else:  # Angular size
+                # Convert angular size to pixels
+                px_size[axis] = (size.to_value(u.arcsec)) / arcsec_per_px
+        else:  # Assume pixels
+            px_size[axis] = size
+
+    if np.all(px_size == 0):
+        # Cross match with point
+        ffi_inds = _crossmatch_point(ra, dec, all_ffis)
+    else:
+        # Cross match with polygon
+        ffi_inds = _crossmatch_polygon(ra, dec, all_ffis, px_size, arcsec_per_px)
 
     return all_ffis[ffi_inds]